Add files via upload
Query starts by taking the fields that will be used in the concatenation to mount the hash manually, including the hashcode that is in the system that is supposed to be wrong. The concatenation function is used within a Hashbytes function for the generation of the MD5 hash, which is the format used by the city in the generation of the hash. After concatenating the query makes the comparison of the concatenated fields that formed the hash with the field that is with the current hash, if there is a divergence the query will return which notes have a problem in the hashcode.